feat: [EL-5188] Participants redesign (collapsed views implementation)
