+++ src/[fsd]/entities/run-history/api/runHistoryApi.js
@@ -1,5 +1,5 @@
 import { RunHistorySerialize } from '@/[fsd]/entities/run-history/lib';
-import { ParticipantEntityTypes } from '@/[fsd]/features/chat/lib/constants/participant.constants';
+import { ParticipantEntityTypes } from '@/[fsd]/features/chat/participants/lib/constants/participant.constants';
 import { eliteaApi } from '@/api/eliteaApi.js';
 import { PAGE_SIZE } from '@/common/constants';
