+++ src/[fsd]/features/chat/participants/lib/constants/index.js
