+++ src/[fsd]/features/chat/participants/lib/constants/participant.constants.js
