+++ src/[fsd]/features/chat/participants/lib/helpers/index.js
@@ -1,0 +1,2 @@
+export * from './addParticipants.helpers';
+export * from './participants.helpers';
