+++ src/[fsd]/features/chat/participants/lib/hooks/index.js
@@ -1,0 +1,4 @@
+export * from './useParticipantEntityIcon.hooks';
+export * from './useAddNewParticipants.hooks';
+export * from './useFetchParticipantDetails.hooks';
+export * from './useParticipantName.hooks';
