+++ src/[fsd]/features/chat/participants/ui/index.js
@@ -1,0 +1,1 @@
+export { default as ParticipantsWrapper } from './ParticipantsWrapper';
