+++ src/[fsd]/features/chat/ui/chat-box/ChatMessageWrapper.jsx
@@ -1,9 +1,9 @@
 import { memo, useCallback, useMemo } from 'react';
 
 import * as ChatHelpers from '@/[fsd]/features/chat/lib/helpers/chat.helpers';
+import { isParticipantStillActive } from '@/[fsd]/features/chat/participants/lib/helpers';
 import { ApplicationAnswer, UserMessage } from '@/[fsd]/features/chat/ui/chat-box';
 import { ROLES, WELCOME_MESSAGE_ID } from '@/common/constants';
-import { isParticipantStillActive } from '@/hooks/chat/useParticipantName';
 
 const ChatMessageWrapper = memo(props => {
   const {
