+++ src/[fsd]/features/toolkits/lib/helpers/toolkitConversation.helpers.js
@@ -1,4 +1,4 @@
-import { ParticipantEntityTypes } from '@/[fsd]/features/chat/lib/constants/participant.constants';
+import { ParticipantEntityTypes } from '@/[fsd]/features/chat/participants/lib/constants/participant.constants';
 import { DEFAULT_MAX_TOKENS, DEFAULT_TEMPERATURE } from '@/[fsd]/shared/lib/constants/llmSettings.constants';
 import { ChatParticipantType } from '@/common/constants';
