+++ src/pages/NewChat/Recommendations/SearchResultList.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useRef } from 'react';
 
+import { getChatParticipantUniqueId } from '@/[fsd]/features/chat/participants/lib/helpers';
 import { ChatParticipantType, PAGE_SIZE } from '@/common/constants';
-import { getChatParticipantUniqueId } from '@/common/utils';
 import useParticipants from '@/hooks/chat/useParticipants';
 
 import NewParticipantList from './NewParticipantList';
